+++ src/util.ts
@@ -17,7 +17,7 @@ export function startedInDebugMode(process: NodeJS.Process): boolean {
   return false;
 }
 
-export const extension = vscode.extensions.getExtension('SonarSource.sonarlint-vscode');
+export const extension = vscode.extensions.getExtension('litesolutions-es.objectscriptquality-vscode');
 export const packageJson = extension.packageJSON;
 
 export let extensionPath: string;